Refactor GUI app entry and update build for JavaFX
Moved JavaFX application logic from Launcher to new GuiApp class for proper Application subclassing. Updated build.gradle to improve JavaFX dependency management, support cross-platform fat JARs, and add verification/distribution tasks. Launcher now delegates to GuiApp.main for cleaner separation.
